Implement a filepaths.h library for handling file paths.
The library knows how to work with path representing files that exist on the
filesystem and path for files that don't exist (yet). The API is meant to be
convenient and cover the classic use cases, that unfortunately are not part
of libc and not always easy to find a reference implementation for that's
ready to incorporate into your own C project.