Implement a run_program API.
The idea is to simply run a subprogram with using fork() and exec() and get
the output of the command in a string buffer. Of course, we need to
dynamically allocate strings to make that happen, and for that we use the
Simple Dynamic String library, another piece of single-header C file lib.