Normalize ../ in filenames, and more clean-up.
It is quite practical to build pathnames using ../ and join operators, and
to make that work seamlessly we need to normalize those "updir"
instructions. The interesting use-case behind that is:

	Path *path = filepath_new("/path/to/some/where/");
	Path *backup = filepath_join_subdir(path, "../backup");

        filepath_ensure_directories_exist(backup);

We certainly don't want to create the "where" directory only so that we can
then move up to the "some" place and create "backup" there.

In passing, also finish our switch to PQExpBuffer.